flask-app/requirements.txt
Flask==2.3.2
psycopg2-binary
flask-cors
